Rails-Server converts params to snakecase
Until now, the keys of the parameters that had to be given to the API
needed to be snakecase because thats what Rails internally uses.
This lead to annoyances in the Client-JS-Code, since we had some objects
that needed to be snakecase to be API-conform.
This works around that by allowing camelCased params for the API.