Split Users name, revamp API, Add Performances
The users name was one field, which could be problematic for a future
feature request like replacing the givenName with "Herr" or "Frau".

The API now always returns the data within a root-object, so that stuff
like metadata and more can be added to the response later on.

In the SegmentPerformances it is now possible to add Performances with a
Performer.